pyproject.toml
[project]
name = "mcpo"
version = "0.0.12"
description = "A simple, secure MCP-to-OpenAPI proxy server"
authors = [
    { name = "Timothy Jaeryang Baek", email = "tim@openwebui.com" }
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "fastapi>=0.115.12",
    "mcp>=1.6.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.1",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.1.0",
    "typer>=0.15.2",
    "uvicorn>=0.34.0",
]

[project.scripts]
mcpo = "mcpo:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
